+++ pkg/application/types.go
@@ -1,8 +1,7 @@
 package application
 
 // ClusterValues holds common values for cluster-<provider> charts. These are
-//
-//	the provider independent values and are present for all the charts
+// the provider independent values and are present for all the charts
 type ClusterValues struct {
 	ControlPlane ControlPlane        `yaml:"controlPlane"`
 	NodePools    map[string]NodePool `yaml:"nodePools"`
@@ -16,4 +15,10 @@ type NodePool struct {
 	Replicas int `yaml:"replicas"`
 	MaxSize  int `yaml:"maxSize"`
 	MinSize  int `yaml:"minSize"`
+}
+
+// DefaultAppsValues holds common values for default-apps-<provider> charts. These are
+// the provider independent values and are present for all the charts
+type DefaultAppsValues struct {
+	BaseDomain string `yaml:"baseDomain"`
 }