+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.9] - 2023-04-26
 
+### Added
+
+- Add `Consistently` function. This takes in a function that returns an error and 
+  runs it for a specified period, stopping on the first error.
+
 ## [0.0.8] - 2023-04-13
 
 ### Added