+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Add DefaultAppsValues
+- Save *rest.Config instead of raw config in the kubernetes client wrapper
+
 ## [0.0.11] - 2023-05-11
 
 ### Changed