+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.12] - 2023-05-16
+
 ### Changed
 
 - Add DefaultAppsValues
@@ -93,7 +95,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added initial framework layout
 - Added Kubernetes client extended from controller-runtime client
 
-[Unreleased]: https://github.com/giantswarm/clustertest/compare/v0.0.11...HEAD
+[Unreleased]: https://github.com/giantswarm/clustertest/compare/v0.0.12...HEAD
+[0.0.12]: https://github.com/giantswarm/clustertest/compare/v0.0.11...v0.0.12
 [0.0.11]: https://github.com/giantswarm/clustertest/compare/v0.0.10...v0.0.11
 [0.0.10]: https://github.com/giantswarm/clustertest/compare/v0.0.9...v0.0.10
 [0.0.9]: https://github.com/giantswarm/clustertest/compare/v0.0.8...v0.0.9